Change name and refactor file structure
