+++ lib/cli.js
@@ -6,7 +6,8 @@ const chalk = require('chalk')
 const indentString = require('indent-string')
 
 const createTestIterator = require('./createTestIterator')
-const { walkSteps, isStepExists } = require('./stepsTraverseUtils')
+const walkSteps = require('./walkSteps')
+const isStepExist = require('./isStepExist')
 const prettyFormatStep = require('./prettyFormatStep')
 const singleton = require('./singleton')
 
@@ -166,7 +167,7 @@ function runDevelopmentMode (testModulePath) {
   .description('Run from current step to step number')
   .action(function (args, callback) {
     handleRun(co(function * () {
-      if (!isStepExists(tester.getTest(), String(args.stepNumber))) {
+      if (!isStepExist(tester.getTest(), String(args.stepNumber))) {
         console.log(chalk.red('Error: step number is not exists in test'))
         return
       }
