+++ lib/isStepExist.spec.js
@@ -1,12 +1,12 @@
 /* eslint-env jest */
 const loadTestModule = require('./loadTestModule')
-const StepTraverseUtils = require('./stepsTraverseUtils')
+const isStepExist = require('./isStepExist')
 
 function load (testModule) {
   return loadTestModule(testModule, { logger: false })
 }
 
-describe('isStepExists', () => {
+describe('isStepExist', () => {
   const test = load(({ step, action }) => {
     step('Step 1', () => {
       step('step 1.1', () => {
@@ -20,9 +20,9 @@ describe('isStepExists', () => {
     })
   })
   it('should be able to determined if steps exists', () => {
-    expect(StepTraverseUtils.isStepExists(test, '1.1')).toEqual(true)
+    expect(isStepExist(test, '1.1')).toEqual(true)
   })
   it('should be able to determined if steps not exists', () => {
-    expect(StepTraverseUtils.isStepExists(test, '1.2')).toEqual(false)
+    expect(isStepExist(test, '1.2')).toEqual(false)
   })
 })
