+++ lib/loadTestModule.test.js
@@ -1,6 +1,6 @@
 /* eslint-env jest */
 const loadTestModule = require('./loadTestModule')
-const { walkSteps } = require('./stepsTraverseUtils')
+const walkSteps = require('./walkSteps')
 
 function load (testModule) {
   return loadTestModule(testModule, { logger: false })
