+++ lib/walkSteps.js
@@ -6,15 +6,4 @@ function walkSteps (root, visit) {
   for (const child of root.children) traverse(child)
 }
 
-function isStepExists (root, number) {
-  let isExists = false
-  walkSteps(root, (node) => {
-    if (node.number === number) isExists = true
-  })
-  return isExists
-}
-
-module.exports = {
-  walkSteps,
-  isStepExists
-}
+module.exports = walkSteps
